enable/disable trace in SmartDashboard
